% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{add_input_process}
\alias{add_input_process}
\title{Add input process}
\usage{
add_input_process(inputid, processid, arg, newfeedname)
}
\arguments{
\item{inputid}{Input to add the process step}

\item{processid}{ID of the processing type to add}

\item{arg}{Arguments for the processing feed}

\item{newfeedname}{New feed name}
}
\value{
An httr response object
}
\description{
Add a processing step to a given input. The processing step will be
placed after any existing process steps.
}
