% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{create_feed}
\alias{create_feed}
\title{Create a new feed.}
\usage{
create_feed(name, tag, datatype = 1, engine = 5, interval = 10)
}
\arguments{
\item{name}{Descriptive name for the new feed.}

\item{tag}{Tag for the new feed.}

\item{datatype}{Defaults to 1 (REALTIME)}

\item{engine}{Timeseries storage engine. Defaults to 5 (PHPFINA)}

\item{interval}{Interval of time series in seconds. Defaults to 10.}
}
\value{
Tibble with id of new feed
}
\description{
Create a new feed.
}
