% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{delete_feed}
\alias{delete_feed}
\title{Delete a feed identified by ID.}
\usage{
delete_feed(feedid)
}
\arguments{
\item{id}{ID (integer) of the feed to delete.}
}
\value{
A tibble of data feed values
}
\description{
Delete a feed identified by ID.
}
