% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{delete_input}
\alias{delete_input}
\title{Delete an input by id}
\usage{
delete_input(inputid)
}
\arguments{
\item{inputid}{ID of the input to delete}
}
\value{
Boolean success/failure
}
\description{
Delete an input by id
}
