% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{delete_input_process}
\alias{delete_input_process}
\title{Delete input process}
\usage{
delete_input_process(inputid, processid)
}
\arguments{
\item{inputid}{ID of feed to modify.}

\item{processid}{ID of the processing step to remove}
}
\value{
An httr response object
}
\description{
Remove a processing step from an input.
}
