% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{emoncms_api_key}
\alias{emoncms_api_key}
\title{Fetch the emonCMS API key from the environment.}
\usage{
emoncms_api_key()
}
\value{
Returns the API key to use for the emonCMS server.
}
\description{
Fetch the emonCMS API key from the environment.
}
