% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_buffer_size}
\alias{get_buffer_size}
\title{Retrieve the number of buffer points pending write}
\usage{
get_buffer_size()
}
\value{
Number (integer) of data points pending flush to disk
}
\description{
Checks the number of cached data feed points in the emonCMS cache layer
(typically Redis) that have not yet been saved to disk.
}
