% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feed_data}
\alias{get_feed_data}
\title{Determine fonts we can/should use.}
\usage{
get_feed_data(id, start = as.integer(lubridate::now() - lubridate::ddays(7)),
  end = as.integer(lubridate::now()), interval = 60 * 30)
}
\arguments{
\item{id}{ID of feed}

\item{start}{Start time (must be coercable by lubridate to unixtime)}
}
\value{
Tibble
}
\description{
Determine fonts we can/should use.
}
