% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feed_fields}
\alias{get_feed_fields}
\title{Get the fields associated with a feed}
\usage{
get_feed_fields(feedid)
}
\arguments{
\item{feedid}{ID of the feed}
}
\value{
A tibble with all feed information
}
\description{
This seems to redundent with the get_feeds function
}
