% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feed_size}
\alias{get_feed_size}
\title{Retrieve the current consumed size of all feeds}
\usage{
get_feed_size()
}
\value{
Integer of bytes consumed
}
\description{
Retrieve the current consumed size of all feeds
}
