% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_input_processes}
\alias{get_input_processes}
\title{Get input processes}
\usage{
get_input_processes(inputid)
}
\arguments{
\item{inputid}{Input ID to get process list.}
}
\value{
An httr response object
}
\description{
Get all of the processes associated with a given input.
}
