% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{list_feeds}
\alias{list_feeds}
\title{Functions for interacting with the emonCMS feeds.
List available feeds}
\usage{
list_feeds(authenticated = TRUE)
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
Functions for interacting with the emonCMS feeds.
List available feeds
}
