% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{list_input_processes}
\alias{list_input_processes}
\title{List input processes}
\usage{
list_input_processes(inputid)
}
\arguments{
\item{inputid}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
List all of the processes associated with a given input.
}
