% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{list_inputs}
\alias{list_inputs}
\title{Functions for interacting with the emonCMS inputs.
List all available inputs and their most recent values.}
\usage{
list_inputs()
}
\value{
A tibble of feeds and their current values
}
\description{
Functions for interacting with the emonCMS inputs.
List all available inputs and their most recent values.
}
