% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{move_input_process}
\alias{move_input_process}
\title{Move input process}
\usage{
move_input_process(inputid, processid, moveby)
}
\arguments{
\item{inputid}{Retrieve authenticated feeds}

\item{processid}{Retrieve authenticated feeds}

\item{moveby}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
Moves the rank order of a given feed's process step
}
