% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{post_bulk_data_to_input}
\alias{post_bulk_data_to_input}
\title{Post data to an input}
\usage{
post_bulk_data_to_input(value, nodeid = "emoncmsr",
  timestamp = as.integer(Sys.time()))
}
\arguments{
\item{value}{Data value to post}

\item{nodeid}{Node ID}

\item{timestamp}{UNIX timestamp}
}
\value{
An httr response object
}
\description{
Post data to an input
}
