% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{post_data_to_input}
\alias{post_data_to_input}
\title{Post data to an input}
\usage{
post_data_to_input(values, nodeid = "emoncmsr")
}
\arguments{
\item{values}{List of name/value data pairs to post}

\item{node}{Node name to post}
}
\value{
Success/failure as a boolean
}
\description{
Post data to an input
}
