% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{read_feed_file}
\alias{read_feed_file}
\title{Read a binary PHPFINA file}
\usage{
read_feed_file(feed_id)
}
\arguments{
\item{file}{Full path to bare (no file extension) feed}
}
\value{
XTS object of the data in the feed.
}
\description{
Read a binary PHPFINA file
}
