% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{reset_input_processes}
\alias{reset_input_processes}
\title{Reset input processes}
\usage{
reset_input_processes(inputid)
}
\arguments{
\item{inputid}{Input ID to reset.}
}
\value{
An httr response object
}
\description{
Removes all processing steps associated with an input
}
