% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{send_emon_request}
\alias{send_emon_request}
\title{Send a request to emonCMS}
\usage{
send_emon_request(uri, params = NULL, verbose = FALSE)
}
\value{
An httr response object
}
\description{
Send a request to emonCMS
}
