% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{set_input_field}
\alias{set_input_field}
\title{Set a field on an input}
\usage{
set_input_field(inputid, field, value)
}
\arguments{
\item{inputid}{ID of the input to delete}

\item{field}{Name of field to modify}

\item{value}{New value of the field}
}
\value{
Tibble with succes, message columns
}
\description{
Set a field on an input
}
