% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{set_input_process}
\alias{set_input_process}
\title{Set input process}
\usage{
set_input_process(inputid, processlist)
}
\arguments{
\item{inputid}{Input ID to modify}

\item{processlist}{Full process list}
}
\value{
A tibble with success/failure and any additional messages
}
\description{
Set the entire processing chain for an input. This appears to be the counter-
part to the `process/add` call from EmonCMS.ORG
}
