% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{add_input_process}
\alias{add_input_process}
\title{List available feeds}
\usage{
add_input_process(inputid, processid, argv, newfeedname)
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
List available feeds
}
