% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{create_feed}
\alias{create_feed}
\title{Create a new feed.}
\usage{
create_feed(name, tag, datatype = c("realtime", "daily"),
  engine = c("phpfina", "virtual"), interval = 10)
}
\arguments{
\item{name}{Descriptive name for the new feed.}

\item{tag}{Tag for the new feed.}

\item{datatype}{Defaults to REALTIME (1). May also be DAILY (2).}

\item{engine}{Timeseries storage engine. Defaults to PHPFINA (5). May
also be VIRTUAL (7)).}

\item{interval}{Interval of time series in seconds. Defaults to 10.}
}
\value{
Tibble with id of new feed
}
\description{
Create a new feed.
}
