% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{delete_input}
\alias{delete_input}
\title{List available feeds}
\usage{
delete_input()
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
List available feeds
}
