% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feed_fields}
\alias{get_feed_fields}
\title{Determine fonts we can/should use.}
\usage{
get_feed_fields(id)
}
\value{
An httr response object
}
\description{
seems to be redundent with the get_feeds function
}
