% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feed_metadata}
\alias{get_feed_metadata}
\title{Gets metadata information for a data feed.}
\usage{
get_feed_metadata(feedid)
}
\arguments{
\item{feedid}{Feed ID to retrieve.}
}
\value{
A tibble of metadata information
}
\description{
Gets metadata information for a data feed.
}
