% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{get_feed_size}
\alias{get_feed_size}
\title{Retrieve the most current value from a given feed.}
\usage{
get_feed_size(id)
}
\arguments{
\item{id}{ID (integer) of the feed to retrieve.}
}
\value{
A tibble of data feed values
}
\description{
Retrieve the most current value from a given feed.
}
