% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{list_input_processes}
\alias{list_input_processes}
\title{List available feeds}
\usage{
list_input_processes(inputid)
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
List available feeds
}
