% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{list_inputs}
\alias{list_inputs}
\title{Functions for interacting with the emonCMS inputs.
List all available inputs}
\usage{
list_inputs()
}
\value{
A tibble of available inputs
}
\description{
Functions for interacting with the emonCMS inputs.
List all available inputs
}
