% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{list_inputs}
\alias{list_inputs}
\title{Functions for interacting with the emonCMS inputs.
List available feeds}
\usage{
list_inputs()
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
Functions for interacting with the emonCMS inputs.
List available feeds
}
