% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{move_input_process}
\alias{move_input_process}
\title{List available feeds}
\usage{
move_input_process(inputid, processid, moveby)
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
List available feeds
}
