% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{move_input_process}
\alias{move_input_process}
\title{Move input process}
\usage{
move_input_process(inputid, processid, moveby)
}
\arguments{
\item{inputid}{Input ID to modify the process list.}

\item{processid}{Which process to modify.}

\item{moveby}{How many slots forward or backward to move the process.}
}
\value{
An httr response object
}
\description{
Moves the rank order of a given feed's process step
}
