% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{post_bulk_data_to_input}
\alias{post_bulk_data_to_input}
\title{Post bulk data to an input}
\usage{
post_bulk_data_to_input(data, reference_time = as.integer(Sys.time()))
}
\arguments{
\item{data}{Dataframe of offset, nodeid, values to post to the input.}

\item{reference_time}{A reference UNIX timestamp to which all offsets are
added/subtracted for each data point. Defaults to the current time.}
}
\value{
Success/failure as a boolean
}
\description{
Post bulk data to an input
}
