% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{post_data_to_input}
\alias{post_data_to_input}
\title{List available feeds}
\usage{
post_data_to_input(id, time, value)
}
\arguments{
\item{authenticated}{Retrieve authenticated feeds}
}
\value{
An httr response object
}
\description{
List available feeds
}
