% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feeds.R
\name{set_feed_field}
\alias{set_feed_field}
\title{Set field values for a feed}
\usage{
set_feed_field(feed_id, field, value)
}
\arguments{
\item{feed_id}{ID of the feed to update}

\item{field}{Name of field to update}

\item{value}{New value of field to set}
}
\value{
An httr reponse object
}
\description{
Note that attempting to set a field to its current value will generate
a failure.
}
